+++ Config.js
@@ -23,10 +23,8 @@
 define(function (require, exports, module) {
 	module.exports = {
 		NAME: 'brackets-minimap',
-		MINIMAP_WIDTH: 220,
 		defaultPreferences: {
-			enabled: true,
-			type: 'codemirror'
+			enabled: true
 		}
 	};
 });