+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.1.0 (22/02/2015)
+* Full code refactoring
+* Plain text mode is removing: excessive feature
+* ADD: minimap is overlay now
+* ADD: minimap width scaling proportional to the width of the editor
+* ADD: Split view support
+
 ## 3.0.1 (18/02/2015)
 * Fixed minimap scroller calculate